Implementation error: incorrect integral manipulation fixed.
Fixed an erroneous dual-input single-compartment liver perfusion model calculation.
I had attempted to shorten a convolution by incrementally summing an integral. It
turns out that this is not a valid simplification because the time step shows up as
an exponential weighting factor within the integral.

It *might* be possible to multiply the previous integrated value by an exponential
to make up for the difference. I need to investigate this further; it could be
numerically unstable or also invalid.