Added support for custom display window centre and level.
The viewer now supports a custom display window and level that is 'sticky' for an
entire Image_Array. Granularity was hard to balance with convenience; ideally you
want something that is non-linear, so it can be adjusted rapidly when you are far
away from the intended values, and be adjusted finely when you are near. My scheme
is not ideal, but lets the centre be adjusted linearly (using the current width as
the scale) and the width be adjusted exponentially. Theoretically, any window
centre and width can be achieved this way. In practice, it is a bit cumbersome
because you might need to iterate adjusting the centre and width back-and-forth a
few times. Works OK for what I needed though.

Also generalized the orthogonal slice generator to handle spiral CT/shuttle mode CT
images, which are spatially sampled in back-and-forth layout instead of the purely
forward layout (i.e., toroidally wrap around from last to first) normally.