Added a dry-run mode for the ingress tool.
You can now use this tool to safely determine if the given file will be permitted into the DB.
The DB will check all row constraints, but will never commit the transaction.
Checking whether the file can be written is not currently tested, but would be welcomed in the future.