Reorganized and renamed much kinetic modeling code.
Trying to make way for implementation of other models and variations on existing models.
Still working to bring the linear interpolation variations up to the same par as the
Levenberg-Marquardt variation of the single compartment, dual-input 5-parameter model
using a Chebyshev polynomial computation method. (The linear interpolation method will
be used for a comparison, but I want to make the comparison as fair as possible, so I'll
try to switch from NLopt 'freeform optimization' to Levenberg-Marquardt least squares
optimization. Work is ongoing.)