Refactored KineticModel_1Compartment2Input_5Param model evaluation routines.
This won't necessarily work for future models or tweaks, because sometimes what is computed
can be reduced based on the optimization method (e.g., if the Jacobian is not used, you don't
bother to compute derivatives -- but if it is, you have to compute it). For this particular
case, I decided it was worthwhile to merge the routines together.

By the way I may have to also overload it based on the computation method, even if the
method is defined separately for different methods. The signature will currently be the same,
but the routines will be different. So the function will end up being defined twice.