Incomplete: switched 1C2I Reduced3Param FreeformOptimization model from NLOpt to G(NU)SL.
The NLOpt optimizers were simply not working for some reason, possibly due to stopping criteria
issues. The problem has been noted before (in this project) and the migration to G(NU)SL fixed
the issue from happening.

Still todo: optimize the optimizer and figure out a way to deal with NaN model parameters correctly.
(Because G(NU)SL does not play nice with NaN's it seems.):