Address a long-standing annoyance in pixel loading code.
I believe the bug is in the Imebra library, but am not sure. It seems that Imebra was not honouring
the choice between signed and unsigned when explicitly asked (as if a reinterpret_cast was being used
but was incorrect) but asking Imebra to convert the number to a double seems to have resolved the
problem.

I suspect that asking for a double results in the correct value because the integer pixel value
cannot simply be reinterpret_cast'ed to a double due to differing sizes. Whatever the case, this
annoyance showed up in MRI and CT data for years. I tried working around it by manually scaling,
looking at DICOM tags and VOI/LUT for transformations, looking at the DICOM PixelRepresentation
directly, relying on Imebra to know the difference via isSigned() all to no avail. Hopefully the
double can always be safely narrowed (e.g., NaN and +-inf), though it shouldn't be an issue at the
moment because I have zero DICOM images with non-integer pixels.