Fixed bug in which clustered surface mesh facets' clusters were not reported correctly.
The problem was that I used the wrong boost::property_map in the reporting code.
(Only two clusters were ever being reported, but it seemed strange.)

I also separated the CGAL SDF computation and normalization stages in anticipation
of possibly using curvature in place of SDF --OR-- manually clustering with raw SDF
values.