Closing theoretically possible bug not encountered.
If the JUNK label is not being used (and so is empty) AND a string garners a
below-threshold score AND the current explicator 'no-match' string is empty,
then a label might be accidentally mapped to the wrong clean label. This
would probably never happen, but fixing it required a single check that the
JUNK label is being used. So why not.