Incomplete: additional enhancements for subsegmentation routine.
Still need to work out the data storage mechanism. I think storing directly into a
database would be ideal, but I'm not sure to let the user specify details. Hard
coding isn't good but is easy for the user; being too flexible means lots of coding
and a brittle implementation.