Added a standalone FITS file loader for easier viewing.
Imagemagick cannot deal with the FITS files I'm creating, but they view quickly in my SFML_Viewer.
A standlone FITS loader means I can inspect pixels and run other analyses on them. Note it is better
to use the Boost.Serialization code when possible because it is more robust and saves all metadata.
FITS are better for interchange, though.