Added a routine for dumping ROI contours.
I needed this for visualization. Currently dumping is a bit ad-hoc because until now Gnuplot was
sufficient to plot the contours. It no longer seems to be sufficient, so I plan to update to
generating a more tailored file. However, there might be a use case for this code down the road.