Surface-based ray casting: only consider a single ray intersection.
Instead of a parametric approach in which I have to locate a plane via centroids,
I now cast rays and consider only the ray nearest the detector. This avoids the
problem of several intersection points for complicated geometry and is essentially
a visibility test. However planes based on centroids are still needed to position
and orient the virtual source and detectors.