Added a git-annex-inspired duplicate cleaner.
It currently relies on adherence to the DICOM standard, in that all modifications to a DICOM
file warrant new unique identifiers. So files that have had small adjustments (e.g., harmonizing
SeriesNumber's, or post-anonymization tag removal) will treated as duplicates if the unique ids
are the same. This is also how the PACS DB will treat them as duplicates or not, but they are not
strictly equivalent for the user in all situations.