Use contour colour to convey contour orientation.
Note that a sampling of Eclipse-produced contours showed that the orientation is NOT uniform,
even in the same dataset. The contour orientations were mostly pointed into the screen contrary
to the normal mathematical positive convention, so I flipped blue and red so that blue shows
more often. Hope this doesn't confuse down the road...