Added ROI exclusion capability.
I've encountered some annoying ROI labels with punctuation that interferes with
(exact) contour recognition. It would be very annoying to rework all of the
Explicator sources to work around this, and I don't think it will be a problem
when non-exact predictions can be used (i.e., in real-world applications), so I
decided it would be best to add a work-around for pruning annoying label names
during iterative lexicon construction.