Capped dose_scaling.
Having a dose scaling factor that dyamically filled the full uint32_t range caused issues with DICOM parsers truncating
digits. I capped it at 1E-5 which is probably sufficient for most physical doses. This may cause issues with
pathological or synthetic doses, but it shouldn't result in any spectacular failures -- just an overall degredation of
precision. Not sure if there is a better solution.