Preliminary support for RTIMAGE files.
The import method relies on the generic CT/MR DICOM image loading routine, which was revamped to allow 'fallback' DICOM
tags to be used. This was required because RTIMAGE files appear to exclusively use their own DICOM tags for image
positioning (understandably). The methods were modeled on the PostGreSQL COALESCE function.

The image loader will now fail instead of creating missing DICOM tags on-the-fly and reporting the default values, which
I think is the default for the (now well-aged) Imebra library version that I am using.

RTIMAGE support is merely preliminary. I haven't tested output in any way. In particular, I don't know whether
Hounsfield units (or some other well-defined units) are supposed to be used.