Web app now focuses on next group box upon interaction.
The existing behaviour was for the view to remain static after pressing buttons or interacting with any elements. Now
the page focus is panned so that newly appearing elements come into view. There are still some issues, such as elements
not being fully loaded when the focus changes, so the scroll appears to be partial when dynamic content is altered.