Added a from-scratch window update routine.
Used it to simplify two Mutate_Voxels calls. I think it will be faster to do a later
pass over the raw voxels rather than visit every voxel one-by-one with a user
functor. It also simplifies the voxel iteration a bit, so win-win.

There is a legitimate scenario in which only altered/selected voxels should be used
for the window-level calculation. This use case is still supported (and is widely
used throughout the rest of the YgorImages_Functors).