Webserver displays generated CSV files.
This is useful because often the CSV files are short and you only care about a few rows/columns. Downloading the file
and opening in another program is overkill for such situations. Table colours were also tweaked to make it easier to
distinguish abutting columns. The colours are 'engineering' colour choices. Output tables are also not sortable, but
there is no reason they couldn't be.