Exposed dose melding operation.
Also had to amend an earlier choice to only honour the last-loaded RTDOSE file.
DICOMautomaton has grown in scope since I made that assumption, and it no longer
makes sense.

I also created a simplistic DroverDebug to help show what the drover object's
state is for debugging purposes. It was valuable for tracking down the source of
the disappearing Dose_Array in this case.