Avoid NaN contamination in LKB NTCP model.
The dose contribution from voxels with zero or near-zero dose to the LKB NTCP model was
causing a NaN to enter the final summation and ruin the model prediction. This patch
suppresses the NaN from such voxels and thus keeps the sum finite.

This approach is appropriate physically since the dose to voxels surrounding voxels
with non-zero dose are almost certainly not exactly zero. By ignoring the contribution,
zero-dose voxels are effectively given a dose of epsilon for some insignificant value
of epsilon.