Light-rad field coincidence analysis operation completed.
Some tweaking might still be needed. It's also cumbersome to have to manually partition the different energies.
It would be nice to be able to accept all the images and perform this separation internally. For the moment, though,
it appears to work well with the test images I've collected so far.