Disallow explicit rescaling for RTDOSE exports.
From what I can tell, leaving these in is within-spec, but causes Imebra to interpret the pixel depth as 1 byte
upon reading, regardless of the on-file pixel depth (i.e., 4 bytes). It could be an Imebra issue, or it could be
contrary to the specification. Either way it appears to not be important for RTDOSE exports.