Picket fence: project analysis results onto isoplane.
Because the DICOM standard gives RTIMAGE pixel spacing at the SID distance (i.e., the hardware pixel spacing), I needed
to transform the results from distances in the SID isoplane to distances in the SAD isoplane. The analysis is performed
on the SID isoplane, so all in-plane SAD isoplane distances supplied by the user are scaled to the SID isoplane.