WebServer: replace Wt animations with pure CSS animations.
There were awkward interactions between the on-the-fly animations being sent by Wt and CSS transitions for image hover
scaling. After a slow opacity darkening there was an abrupt Wt-caused white flash and slow fade-in of the image,
presumably due to stacking animations and transitions. The flash was caused by a parent noticing that a child was
impacted by a :hover event, and I was unable to block the parent from reacting. Thankfully it was straightforward to
implement most of the Wt animations with a few lines of CSS. Tables do not yet properly fade out, but this is a small
nuisance.