Added an operation for static beam optimization.
This implementation focuses on optimizing beam weights to achieve a dose distribution that satisfies a hard constraint
D_max = X and optimizes the number of voxels with dose > Y. Currently there are no target requirements -- the dose is
optimized and the target is assumed to comprise the majority of the input (so it should be cropped beforehand).

Weights are forced to sum to 1 and assumed to be equal when fed to the optimizer. The normalization is arbitrarily
chosen so that the D_max constraint is satisfied.