Better PresentationImage defaults.
The PresentationImage operation lacked (and still lacks) an image-of-interest selector, so
I updated the defaults to display: (1) the last image array in the stack, and (2) the slice
in this array closest to (0,0,0) in PresentationImage operation. This image array and slice
are more likely to be of interest that the first image in the first array.

The default is now to automatically convert dose arrays to image arrays when there are no
image arrays available. This is a more sensible default than throwing an exception, but also
mutates state which may be unexpected for a viewer. A more robust dose+image array viewing
system would be better. Even better will be to reduce the source-level differences between
dose and image arrays...