Added partial support for Boolean negation.
A full implementation will have to split the selection of contours to operate on, and contours to match.
This will avoid the ambiguity of whether to purge a contour that matches the regex and ultimately allow a much greater
control to build complex Boolean statements. This commit is a stop-gap that can accomplish a simple negation of the
provided regex and provide a 'keep iff' operation.