Support for surface mesh processing.
Basic operations for generating, smoothing, simplifying, and writing to file are provided. A routine for slicing meshes
to produce contours is also provided. Mesh generation can be performed with an implicit surface mesher built on
restricted Delauney triangulation using a grid-based approach, or with an advancing-front algorithm using contours
directly.

A generic 3D Minkowski sum operation is built on top of the primitives as a proof-of-concept. It's extremely slow
because it does not take any computational short-cuts and scales badly, will only work on simple ROIs without holes, and
can only reasonable be used with a meshes with a few hundred faces. Sadly because it scales so badly the surface mesh
must be heavily simplified and so the core benefit of an exact implementation is lost. However, it might be useful to
have around to benchmark future algorithms.