Static beam optimizer: switch from hypersphere-constrained parameters to direct hyperplane parameters.
I could not figure out how to inform the optimizer of all the beam weight constraints, so the new method is worse
mathematically since the optimizer is unaware of a constraint that is applied on-the-fly (ensuring beam weights sum to
one). There is a degree of freedom not genuinely being optimized, but the optimizer is unaware and it effectively acts
like a strong correlation between all parameters, so it is possible that some optimizers will be compromised in certain
situations.

The speed appears to be faster, and I think the hyperplane step size should be more uniform across the entire parameter
space compared with the hypersphere approach due to its nearly linear behaviour (except for the rescaling DOF that ruins
it). The hypersphere approach may actually produce better results though. This implementation should be tested on a
variety of real world data before settling.