WIP: CompareImages: added operation for comparing image arrays.
This WIP commit uses Boost R*-trees to attempt lazy nearest-neighbour lookups to handle badly mismatched data. I could
not figure out how to effectively accomplish lazy nearest() queries with Boost, and it would probably be best for
pixel-to-pixel discrepancy to guarantee the pixels overlap within some tolerance. Given the regularity of the data, I
think I can more economically replace the R*-tree with a 3D circulator. However, there are instances in which the
geometry is badly mismatched where a fully generic R*-tree would be nice to have. I haven't decided if it would be
worthwhile to keep or discard yet.