WIP: CompareImages: replace R*-tree approach with grid-traversal.
This commit required a fair amount of support to be developed for 3D traversal of rectilinear grids that was not
previously needed. Some of the functionality in the CompareImages operation may be moved into a Ygor class to
provide tooling for other work (e.g., registration, hit-and-miss operators, full 3D convolution operations, voxel
growth/expansion, etc.).

This implementation also supports test images (i.e., the images that will be edited) that are either aligned or
un-aligned with the reference (rectilinear) image volume grid. Aligned images are special-cased to reduce per-
voxel sampling costs, which were found to be significant in some circumstances.