WIP: DetectGrid3D: display proper grid line contours.
This commit adds instrumentation for digging into the fitting code. At a glance, it seems to work well. The next
approach will be to use synthetic data, or perhaps something that more closely follows a known grid, so that performance
can be understood. This will shape the direction the algorithm is taken.