VolumetricSpatialDerivative: switch to pre-indexed trilinear interpolation.
This halves the speed required for a typical test case, but is still not particularly fast for even a medium-sized image
volume. It could be made extremely fast by manually iterating over the voxels in the nms_working volume and the
imagecoll volume. However, it would be tricky to get right since contours are involved.