ModifyImageMetadata: allow modification of image spatial characteristics.
This change is currently targetted toward single images only. Applying to image arrays will result in overlapping
images, so the images in an image array need to be inconveniently exploded into individual image arrays, or this routine
will need to be modified to accept 'dR'-style vector offsets for multiple images.

I don't yet have a good feel for how to design this, so I'll leave it until it's needed.