WIP: added operation for drawing geometry into images.
A lot of the 'Generate*' operations share a common base of image creation that would be best factored or offloaded to a
dedicated routine. The GenerateSyntheticImages routine effectively already abstracts this aspect, so a complementary
routine for drawing the geometrical shapes and patterns on the images is needed to complement it.

This commit is a WIP that only contains a cylindrical grid drawing routine. Support for more shapes will be added
shortly, ideally re-using the ContourViaGeometry parameter specifications.