Add a volumetric spatial blur operation.
This operation currently only boasts a single, fixed Gaussian option, but this is one of the most versatile and
widely-used blurs. This operation was originally written to support dynamic windowing, but was abandoned due to
difficulties in assessing the DICOM distance to adjacent voxels. Instead, a fixed sigma = one pixel and 3*sigma window
was hard-coded. The blur can be widened via successive applications of the same blur.