FVPicketFence: use whitelist for exposing parameters.
Previously a blacklist was used. This meant that future changes to constituent operations would alter which parameters are
shown to the user, which is not acceptable for a 'simplified' interface. Switching to a whitelist will make these
parameters take their default values, which will (hopefully) be suitable, or at least fail rather than confuse.