BoostSerializeDrover: support export of inidividual components.
This will provide greater flexibility when fusing data from multiple analysis phases, for example contouring without
having to export as an RTSTRUCT or dose without having to export as an RTDOSE file. This can be important since not all
files support all DCMA features, and a DICOM-inconsistent state might be valid (and preferred) in some contexts.