Added runtime ELF patching for portable dcma binary.
This is useful when the versions of glibc differ slightly. There is of course some risk in running binaries this way,
since some expected glibc functionality will be missing. However, this distribution method is mostly needed for one-off
applications that can be manually verified as necessary and debugging on underpowered devices.