Support separate operation-parameter specification.
Some operations with many parameters can result in invocations that are unwieldy. Tweaking invocations can also be
annoying if all parameters are sequential and individual parameters cannot easily be disabled.

This commit enables parameters to be specified not only along with the operation (i.e., inline -- the previous
behaviour), but now also after an operation name is specified. Individual parameters can be toggled as enabled or
disabled. All disabled operations will implicitly disable trailing parameters. Also, parameters can still be specified
in groups for convenience or to denote logical relationships.