Surface_Meshes: specify image traversal order.
Image traversal order is not important for the correctness of the existing algorithm, but may be necessary if
algorithmic simplifications will be applied in a later commit.

An image adjacency lookup was hoisted out of the main loop. This may slightly speed up the algorithm, but the majority
of time is still spent de-deuplicating vertices and constructing the mesh from a polygon soup.