Optional external dependencies: support conditional compilation for major external dependencies.
This commit provides a means of disabling CGAL functionality, which should ease porting considerably if CGAL
functionality can be dropped. Take note that a *lot* of functionality is lost when CGAL support is disabled.
See Operation_Dispatcher.cc for list of operations that are explicitly impacted; also note that some loss of
functionality may only be evident at runtime.