Meshing: prune irrelevant candidate vertices.
All possible vertices were being considered in the mesh construction phase of the marching cubes implementation. In the
worst case scenario (lots of voxels), entirely irrelevant candidate vertices were being considered. This fix addresses
the worst-case scaling issue, but does not fully solve the problem. Note that this fix also requires images to be
ordered (which is enforced, but potentially something of note).