SimulateRadiograph: faster ray marching.
Moved from a 'march-and-interpolate' model to a Bresenham's-like (i,j,k) traversal model to dramatically speed up
simulated radiographs. The implementation could still do with a bit more cleanup, but I'm planning to extract a somewhat
generic ray traversal routine for calling user-provided routines on the voxels intersected during transit.

In addition to the speedup, an alternative image contrast mode is now available that eliminates all physical parameters.
The generic radiograph-style exponential attenuation model is retained, but now an alternative that is effectively the
log of the exponential model (i.e., sum of mu*dL over all intersecting voxels) is available. This model is nice because
it should almost always give a reasonable image, whereas the exponential model often requires some tweaking and can
naively result in images with nearly zero contrast.