SimulateRadiograph: truncate rays to image volume bounding box.
This commit significantly speeds up ray marching. It also supports ray line segments that terminate inside the bounding
volume, though support is somewhat limited since the images are extremely blocky. Also note that radiographs may be
clipped if the ray source is near the volume. When it is far away the effect is minimized.

This commit also includes some (disabled) code to generate an OFF file that displays ray line segments, image volume,
and the image volume centre. It worked well for debugging and verifying the operation, but doesn't seem reaonable enough
to keep so will be removed in the next commit. It is included here for reference only.