File_Loader: support (uncompressed) TAR files.
This commit permits uncompressed TAR files (a subset of ustar and posix/pax variants) to be read. At the moment, if an
encapsulated file cannot be loaded, the TAR file loader will consider it to be a fail.

Since some aspects of TAR files are unsupported, and some of the unsupported features could be interpretted as
encapsulated files (e.g., pax long filename extensions), valid files could in some situations fail to load. While this
might be inconvenient, it is most conservative in limiting the possibility of erroneous data being loaded. Conversion to
the ustar format will provide a workaround if the problems arise from tar format incompatibilities, and elimination of
irrelevant files will permit TAR files to be successfully loaded.