Support reading and writing of gzipped-TAR files.
Currently all encapsulated files are loaded one at a time. This can cause, for example, a logically-grouped CT image
volume of N images to be broken into N separate volumes. It isn't clear how best to handle this, though inclusion of
multiple files into a single TAR archive is at least somewhat suggestive of a logical relation. This aspect may be
addressed in the future, but at the moment images should be explicitly joined together as needed.