Added a rectangular, anisotropic voxel volume, gzipped TAR file test file.
This file contains the DCMA logo and text. It will only be readable if the image orientation is handled correctly.
Additionally, viewers that ignore the pixel width and height will display a stretched image. Finally, the file is
encapsulated in a gzipped (GNU) TAR file to test that such files can be read.