CI test whitespace change.
The CI Docker setup was recently changed, but the compilation failed due to an HTML 429 error.
My xorg-server or logind server died immediately after I launched the CI Docker build locally.
It isn't clear whether the 429 error was caused by something in the CI script, or was a fluke.
This whitespace-only commit will help determine the origin...