Linux: include non-library packaged files for DCMA.
This commit adds packaged component replication for all files listed in the host's package manifest. This only seems to
be needed for DICOMautomaton-specific files, but other core packages could be added as needed.

Also note that the initramfs is growing fairly large, near 1GB. It may not be an issue since most machines are likely to
have at least 4 or 8GB these days, it is a significantly large file. In the future a minimal initramfs may be separated
from the rest of the packages, with the bulk placed on a separate disk image.