Histogram sub-system overhaul.
Reworked how histrograms are computed in order to be significantly faster. Previously all interior voxels were cached and
the cumulative histogram was computed directly. Now, a differential histogram is computed in two passes, and the only
cache is for multiple threads, and is much more limited.

The histogram-generating routine has been made more generic (thus the mass renaming) and now returns a samples_1D
directly rather than voxel value caches. While this is more direct, the downside is that distribution statistics can not
be computed as precisely (e.g., using the YgorStats' Stats::Sum/Mean) but are now computed in an approximate way. They
still use a correction-based summation approach, but are not pre-sorted to minimize numerical losses.