Support 'tabular DVH' formatted text files.
I'm not certain the format is *fully* supported; I couldn't find any documentation and it seems there are concessions
for multiple plans to be reported, and differential histograms. I don't have any file specimens to deal with such
situations yet, but the implementation is relatively compartmentalized so supporting them in the future should be
reasonably straightforward.

File specimens with both relative and absolute axes normalization are included as test files. (These files, and only
these files, were used to create the file loading routine.)