WIP simple meshing routines.
These meshing routines are designed to be built around a few direct contour-to-contour associativity sub-routines. They
are not meant to comprehensively deal with all possible topologies, only simple cases (e.g., spheres and sphere-like
ROIs). They are also meant to be relatively fast and not introduce any additional vertices in the contours, though this
might be relaxed at some point in the future since mesh quality can quickly drop when sampling rates are too low.