Added direct contour->mesh conversion operation.
There are several caveats to such conversions, which are described in the operation's documentation. Some can be fairly
easily addressed via use of a more sophisticated contour-contour registration routine, but some are intrinsic to the
approach and stem from topological issues that seem difficult to address in a simple way. Nevertheless, this routine
quickly meshes contours from convex polyhedral quickly and with minimal extra vertices.