WIP RPM-TPS implementation.
This commit adds preliminary, undocumented support for the RPM-TPS deformable registration algorithm on point clouds.
The implementation does not always provide stable alignments, and requires both algorithmic and 'operational aspect'
tweaking. This implementation, or bits and pieces of it, will likely be hoisted into a different location for easier
re-use.