RPM-TPS: switched underlying TPS solver.
Switched from the formulation described by Chui and Rangarajan (2001) to a much earlier, possibly more generic approach
described by Bookstein (1989). I extended Bookstein's approach with regularization (as described by Chui) and also kept
the 'double-sided outlier handling' described by Yang et al (2011). The soft-assign technique has been updated as per
Chui's description (it previously incorrectly included outlier weights and may even have read out of a std::vector
bounds, though I never noticed any incidences of obviously invalid outputs and didn't notice any change in the outputs
when I initially corrected it).

Note that the current implementation is rather slow. It appears to be more robust than the previous implementation
though. In particular, I directly invert Bookstein's 'main' TPS matrix 'L', but I use the pseudo-inverse to avoid
inversion difficulties. I will investigate the robustness and speed differences of switching to LDLT; I expect it to be
faster but possibly less robust to degeneracy.

Also note that if regularization is not performed then the 'L' matrix only needs to be inverted once. It does not depend
on the correspondence or later refinement in Affine and warp transformation components. If it is known that
regularization is not needed (i.e., that the inputs are full-rank, i.e., not degenerate, and without outliers) then the
implementation can be significantly sped up.

The reason the TPS implementation was wholesale switched is that I simple could not get Chui's algorithm to produce
reasonable output. It is possible that my initial starting temperature was too hot, which caused the moving point cloud
to collapse to the stationary point cloud's centre of mass. But I wasn't able to correct this behaviour with simple
adjustments. Bookstein's approach gives the same outputs (from what I can tell) but immediately worked without having to
troubleshoot.