TPS: promoted transform class.
The TPS class can be treated as a first-class transformation similar to the affine transformation class, meaning it can
be applied to objects, selected, written to file, etc.

The TPS method in ExtractPointsWarp now exposes the kernel dimension, solver type, and regularization parameter to the
user. The TPS-RPM method does not yet expose any parameters (it will shortly), but the resulting TPS is handled as a
first-class transformation type.