TPS-RPM: initial support for hard constraints.
This commit roughs-in support for forced correspondence, i.e., hard constraints. This functionality is important for
providing contextual information to the optimization scheme, such as fixing the boundaries. The final correspondence is
also now optionally reported, though this functionality is predominantly meant to support internal routines at the
moment.

Some useful debugging visualization tools are included. They can help make sense of the final transformation by showing
which points correspond (though only in the binary, T -> 0 limit, which can potentially be misleading). Using these
tools it seems that mirroring and twisting transformations are not being sufficiently penalized, so I plan to implement
the QR decomposition approach afresh to see if the affine component will be more well-behaved.