TPS-RPM: ported reference QR solver.
I ported the earlier QR solver approach, injecting it on to the current data structure for testing purposes. This commit
also includes a number of small subroutines that are useful for inspecting transformations. I might create a new
operation to house them, though I'm not yet sure if I want to lug around correspondence info in the Transform3 class.