TPS-RPM: add tolerance-based Sinkhorn algo break.
The Sinkhorn softassign algorithm now terminates automatically when the specified tolerance is reached. This is more
directly useful than specifying a number of iterations and hoping it is sufficient for the normalization to converge.

The existing Sinkhorn parameter controlling the number of iterations was converted to an upper limit, and if the upper
limit is exceeded without normalization being within tolerance than the TPS-RPM algorithm terminates with a message to
the user suggesting a greater number of iterations. This behaviour *may* need to change in the future, but for now it
seems reasonable.