TPS-RPM: re-implement double-sided outlier handling
This commit re-implements Yang et al.'s "double-sided outlier handling,"  but re-cast in the style of Bookstein's TPS
solution method.

This commit also permits the user to specify, explicitly, outliers in either the moving or stationary point sets. The
interface is a bit clunky, but it does the job.

Handling explicitly-marked outliers (i.e., 'perfect' outliers) when the double-sided approach is used is tricky since
the weighting scheme results in inf weights. I implemented a kludge in this case to adjust infs to 'some large number',
but it doesn't work particularly well. In fact, specifying the wrong point as an outlier can dramatically alter the
energy landscape and, for example, cause the entire moving point set to collapse to a single, degenerate point. Luckily
the original formulation by Chui et al. is not susceptible to this problem, so it can act as a backup option in these
circumstances. The all-around best approach is to purge known outliers before attempting to register point sets.