TPS-RPM: let user disable outlier handling.
This commit adds the ability to selectively disable outlier detection and handling in the moving point set, stationary
point set, or both.

One limitation of the Sinkhorn technique seems to be that the larger point set has to leave outlier handling enabled.
Otherwise the Sinkhorn technique fails to converge. This behaviour is consistent when one switches the moving and
stationary sets. I also tried to suppress, rather than outright disallow outlier detection in this case, but it failed
to work. The problem seems to be a tied to the method rather than to the implementation.

Interestingly, if both point sets have the same number of points then the Sinkhorn method should still converge as
usual. So there is a workaround if both point sets can be sampled appropriately.