Add minimal test setup.
This test setup will eventually be filled in and may be integrated into the build process. At the moment I'm struggling
to identify a good way to combine unit tests, integration tests, and examples (that can be tested without too much
obfuscation). For the moment I'll rely on a unit testing framework, but may have to either shell out or build a custom
test system; neither would be ideal.