TPS-RPM: report bending energy.
I'm not 100% certain this bending energy is valid in all cases, but I suspect it is. The definition comes from
Bookstein. Also, the bending energy is not the full system energy, but it is something that can be compared when
different regularization parameters and starting/ending/step temperatures are used.