ReduceNeighbourhood: support adaptive normalization.
This commit adds a locally-adaptive non-parametric voxel value normalization technique based on percentiles. The
neighbourhood surrounding each voxel is scanned, sorted, and the percentile of the central voxel value is identified.
The voxel then takes on the value of the percentile.

I also added a simple integration test and a note in the NormalizePixels routine that discusses where to look for
adaptive algorithms.